requirements.txt
pandas
Pillow
pre-commit
termcolor
torch
torchtext
torchvision
